Use relative links setting
When we deploy the site to Github pages, internal links don't work, and links
to stylesheets are broken.

One workaround is to use the Middleman setting `http_prefix` and set this to
the name of the repo, but this breaks some things for example the Tech Docs nav
bar.

Tech Docs have a setting we can use to make it use relative links, which works
both locally and on Github Pages, documented here:
https://tdt-documentation.london.cloudapps.digital/configure_project/global_configuration/#source-url